fix: ensure analytics landing page compiles
